Added comment section in HTML to put in the readme later showing how to use the app among many changes
Completely changed how the app works so it no longer relies on ID's and instead works based off classes. Meaning new accordions can be added with ease by simply copy and pasting the HTML portion in. If wanting to add new accordions with different information in it the process is explained in the README or comments. Need to make even more dynamic for different templates!